requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
bleach==6.2.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
exceptiongroup==1.3.0
fastapi==0.118.0
filelock==3.13.1
fonttools==4.59.2
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.4
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kaggle==1.7.4.5
kiwisolver==1.4.7
labelImg==1.8.6
lxml==6.0.2
MarkupSafe==2.1.5
matplotlib==3.9.4
mpmath==1.3.0
narwhals==2.5.0
networkx==3.2.1
numpy==2.0.2
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.2
pillow==11.0.0
polars==1.33.1
protobuf==6.32.1
psutil==7.0.0
pyarrow==21.0.0
pydantic==2.11.9
pydantic_core==2.33.2
pydeck==0.9.1
pyparsing==3.2.4
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.48.0
streamlit==1.49.1
sympy==1.13.1
tenacity==9.1.2
text-unidecode==1.3
threadpoolctl==3.6.0
toml==0.10.2
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tornado==6.5.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
ultralytics==8.3.199
ultralytics-thop==2.0.17
urllib3==2.5.0
uvicorn==0.37.0
watchdog==6.0.0
webencodings==0.5.1
zipp==3.23.0
